chore: 서비스 내 모든 꿀 타입 통일 (기존 Integer -> Long으로 수정)

+++ match-service/src/main/java/com/lgcns/bebee/match/application/usecase/GetAgreementUseCase.java
@@ -60,8 +60,8 @@ public class GetAgreementUseCase implements UseCase<GetAgreementUseCase.Param, G
         private final LocalDate endDate;
         private final List<ScheduleDTO> schedules;
 
-        private final Integer unitHoney;
-        private final Integer totalHoney;
+        private final Long unitHoney;
+        private final Long totalHoney;
 
         private final Long otherId;
         private final String otherProfileImageUrl;
